+++ Engine/src/Platform/GLFW/GLFWInput.hpp
@@ -14,8 +14,8 @@ class GLFWInput : public Input
   protected:
     virtual bool                    IsKeyPressedImpl(int keycode) override;
     virtual bool                    IsMouseButtonPressedImpl(int button) override;
-    virtual std::pair<float, float> GetMousePositionImpl() override;
-    virtual float                   GetMouseXImpl() override;
-    virtual float                   GetMouseYImpl() override;
+    virtual std::pair<double, double> GetMousePositionImpl() override;
+    virtual double                  GetMouseXImpl() override;
+    virtual double                  GetMouseYImpl() override;
 };
 } // namespace Engine
